requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.7
idna==3.3
Pillow==9.0.0
praw==7.5.0
prawcore==2.3.0
pygame==2.1.2
python-dotenv==0.19.2
requests==2.26.0
update-checker==0.18.0
urllib3==1.26.7
watchdog==2.1.6
websocket-client==1.2.3
